requirements.txt
Flask==0.12
algoliasearch==1.14.1
python-dotenv==0.7.1